pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0.post1"
pydantic-settings = "^2.1.0"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.5.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
